+++ src/main/java/lb/ferzshow/repository/RiderRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.transaction.annotation.Transactional;
 
 @Transactional(readOnly = true)
-public interface RiderRepository extends JpaRepository<Rider, Integer> {
+public interface RiderRepository extends JpaRepository<Rider, Long> {
 
 }
